+++ client/src/components/common/mainTitle/MainTitle.tsx
@@ -1,0 +1,15 @@
+import { MainTitleContainer, MainTitleHeader } from './Style';
+
+interface Props {
+  title: string;
+}
+
+const MainTitle = ({ title }: Props) => {
+  return (
+    <MainTitleContainer>
+      <MainTitleHeader>{title}</MainTitleHeader>
+    </MainTitleContainer>
+  );
+};
+
+export default MainTitle;
